Consolidate keyvault and cryptography client options and model files into single headers. (#2640)
* Consolidate keyvault options and model files into single headers.

* Consolidate cryptography client model headers into a single header.

* Fix up merge conflict

* Consolidate LROs into one, and fixup PagedResponse name.

* Fix up clang formatting and match cpp file name with hpp.

* Fix clang formatting.

* Move LROs and PagedResponse types into models as well.

* Move encrypt and decrypt parameter types into cryptography model header
file.

* Fix clang format, reodering of headers.

* Update to match recent changes.

* Update CL
