+++ sdk/keyvault/azure-security-keyvault-keys/src/import_key_options.cpp
@@ -4,7 +4,7 @@
 #include <azure/core/internal/json/json.hpp>
 #include <azure/core/internal/json/json_optional.hpp>
 
-#include "azure/keyvault/keys/import_key_options.hpp"
+#include "azure/keyvault/keys/key_client_options.hpp"
 #include "private/key_constants.hpp"
 #include "private/key_serializers.hpp"
