+++ sdk/keyvault/azure-security-keyvault-keys/src/key_curve_name.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/keys/key_curve_name.hpp"
+#include "azure/keyvault/keys/key_client_models.hpp"
 #include "private/key_constants.hpp"
 
 namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
